backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
google-generativeai
playwright
python-multipart
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
